adds logic to evaluate user type and render appropriate route

+++ src/components/auth/Login.js
@@ -8,6 +8,7 @@ export const Login = props => {
     const password = useRef()
     const existDialog = useRef()
     const passwordDialog = useRef()
+    const userType = useRef
 
     const existingUserCheck = () => {
         // If your json-server URL is different, please change it below!
@@ -21,10 +22,14 @@ export const Login = props => {
 
         existingUserCheck()
             .then(exists => {
-                if (exists && exists.password === password.current.value) {
+                if (exists && exists.password === password.current.value && exists.userType === 1) {
                     // The user id is saved under the key app_user_id in local Storage. Change below if needed!
                     localStorage.setItem("app_user_id", exists.id)
                     props.history.push("/")
+                } else if (exists && exists.password === password.current.value && exists.userType === 2) {
+                    // The user id is saved under the key app_user_id in local Storage. Change below if needed!
+                    localStorage.setItem("app_user_id", exists.id)
+                    props.history.push("/dashboards")
                 } else if (exists && exists.password !== password.current.value) {
                     passwordDialog.current.showModal()
                 } else if (!exists) {
@@ -45,7 +50,7 @@ export const Login = props => {
             </dialog>
             <section>
                 <form className="form--login" onSubmit={handleLogin}>
-                    <h1>Application Name</h1>
+                    <h1>Bespoke Hub</h1>
                     <h2>Please sign in</h2>
                     <fieldset>
                         <label htmlFor="inputEmail"> Email address </label>
@@ -70,9 +75,9 @@ export const Login = props => {
                     </fieldset>
                 </form>
             </section>
-            <section className="link--register">
+            {/* <section className="link--register">
                 <Link to="/register">Not a member yet?</Link>
-            </section>
+            </section> */}
         </main>
     )
 }
